Condense question rephrasing into single class
Now we've gotten rid of the OpenAI question rephraser, we can simply
the code by condensing the question rephrasing into a single class.
