+++ lib/answer_composition/composer.rb
@@ -38,7 +38,7 @@ module AnswerComposition
       when "claude_structured_answer"
         PipelineRunner.call(question:, pipeline: [
           Pipeline::JailbreakGuardrails.new(llm_provider: :claude),
-          Pipeline::QuestionRephraser.new(llm_provider: :claude),
+          Pipeline::QuestionRephraser,
           Pipeline::Claude::QuestionRouter,
           Pipeline::QuestionRoutingGuardrails.new(llm_provider: :claude),
           Pipeline::SearchResultFetcher,
